add simplified version of DB creation
Created a readable script that resembles the method shown in class.
Fix sensor not having a composite primary key.
`db_create.sql` seems more complete as the foreign keys are described as a constraint and can trigger actions.
Collation information is useful but not crucial.
Every column is forced to NOT NULL even though this is not specified.