upgrade table generation
The table generator now accepts a displaying title for the head of the table, the corresponding data column to represent and special formatting that allows us to create dynamic text (or urls for instance) using the content of any column.
For example,  '<a href="patient.php?id=$number">$name</a>' gets the patient id and name and generates a pretty url in the table itself.